Make ibus-typing-booster multilingual, i.e. use more than one language per engine
The internals now support an unlimited number of transliterations
and dictionaries.

If one wants to test using more transliterations and dictionaries at the
same time one can set dconf keys for example like this:

    dconf write /desktop/ibus/engine/typing-booster/typing-booster-de-de/dictionary '"mr_IN,en_GB,de_DE"'

    dconf write /desktop/ibus/engine/typing-booster/typing-booster-de-de/inputmethod '"mr-itrans,NoIme,t-latn-post"'

The first in this list of input methods is used for the preëdit,
candidates are calculated from all of them. To change the input method
used for the preëdit, the keys Control+ArrowDown or Control+ArrowUp
are used.