Minimal change for the setup tool to enable multilingual support
The internals of ibus-typing-booster now support an unlimited number
of transliterations and dictionaries. But to support this fully in the
setup tool, a complete redesign of the setup tool seems necessary.
For the moment I just add a simple checkbox “Add direct input” which
enables direct keyboard input in addition to the selected
transliteration(s) and adds the en_GB dictionary. I.e. if this option
is enabled, one usually has two input methods and two dictionaries
(unless the selected input method was direct input already and
the dictionary was already en_GB).

This is good enough for the moment to enable something like
“any Indian language + English”.

If one wants to test more transliterations and dictionaries at the
same time one can set the dconf keys directly.

If more languages and dictionaries turn out to be useful, I’ll improve
the setup tool.