The combobox in the setup tool should show the first supported ime from dconf
Before this fix, it showed the first ime from dconf which was not 'NoIme'.
Therefore, if for the de_DE engine the dconf key was

   inputmethod='NoIme,t-latn-post'

the combobox incorrectly showed 't-latn-post' as the selected input
method although 'NoIme' is also a officially supported input method
for the de_DE engine.

And, if for the mr_IN engine the dconf key was

   inputmethod='t-latn-post,NoIme,mr-itrans'

the combobox incorrectly showed mr-inscript2 as the selected input
method because 't-latn-post' was the first input method in dconf which
was not 'NoIme' and then 't-latn-post' was found as not supported by
mr-inscript2 so it showed the first of the supported input methods
which was mr-inscript2.  Now, if finds that mr-itrans is the first
input method from dconf officially supported by the mr_IN engine and
the combobox displays 'mr-itrans', which seems more correct in this
case.