Make it possible to enter a space into the preëdit by typing “G- ” (AltGr+Space)
Sometimes it is useful to enter spaces into the preëdit, for example
to get words commonly used together like “going to” committed as one
unit. Or to help with the matching of emoji (“smiling face” is easier
to match than “smilingface”).  Therefore, there needs to be a way to
prevent a space from committing the preëdit and insert a space
instead. “G- ” (AltGr+Space) seems to be a reasonable choice, it does
not seem to conflict with other stuff (“C- ” would conflict with
“set-mark-command” in Emacs for example).