If the query string in EmojiMatcher.candidates() is an emoji itself, match similar ones
When using backspace to go back to a word, the word is put into
preëdit again and candidates are displayed again to be able to make
corrections. If the “word” reached by backspace was an emoji, nothing
happened because an emoji in the preëdit didn’t match any candidates,
emojis were only matched by “real” text. This is not so useful,
this change puts related emoji in the candidate list if the
preëdit is already an emoji. This is more helpful if one wants to
correct the last typed emoji.

Works only when surrounding text is supported of course.