Remove self._typed_string_when_update_candidates_was_last_called variable
This was a hack because update_candidates was called to often for no
reason from various places, for example when only the cursor within
the preëdit string had moved.  The expensive function
update_candidates() should not be called when it is not necessary.

So this commit fixes the problem of the unnessary calls of
update_candidates() and removes that
self._typed_string_when_update_candidates_was_last_called variable.