Fix bug when committing the preëdit with Space when no candidates are available
    self.forward_key_event(key.val, key.code, key.state)
    return True

needs to be used instead of just

    return False

because the latter works correctly in GTK applications but not in Qt
or X11 applications. When “return False” is used, the key event which
triggered the commit here arrives in Qt or X11 *before* the committed
string. I.e. when typing “word ” the space which triggered the commit
gets to application first and the applications receives “ word”.  See:
https://bugzilla.redhat.com/show_bug.cgi?id=1291238

I fixed this already in another place, but forgot the corner
case when no candidates are available. This corner case has become
much more common now because of the speed optimization not to
calculate the candidates when the lookup table is hidden anyway.

This speed optimization was introduced in the last commit which changed
the behaviour of the Tab key in the mode when “☑ Enable suggestions by
Tab key” is checked