Shortcut keys which look up related candidates should enable the candidate list
Looking up related candidates can currently be called using the
shortcut key AltGr+F12. When the shortcut key is used while no
candidate is selected, it requests to lookup related candidates for
the contents of the preëdit.  This should also work when suggestions
are only enabled by Tab and are currently disabled.