Changes in itb_emoji.py necessary because of the update of the CLDR emoji annotations
After updating the CLDR emoji annotations, some emoji have keywords only
but lack names (type="tts") in some languages. For example, 🤔 has no name
in fr.xml. Here are the lines around 🤔 in fr.xml:

  <!-- 1F607 smiling face with halo; angel; face; fairy tale; fantasy; halo; innocent; smile -->
  <annotation cp="😇">ange</annotation>
  <annotation cp="😇" type="tts">auréole</annotation>
  <annotation cp="🤔">visage | réflexion</annotation>
  <!-- 1F610 neutral face; deadpan; face; neutral -->

I don’t know whether that is intentional or a bug in the CLDR annotations.

But whether it is a bug or nor, it is a good idea to make my code
handle this case reasonably as well.