Fix behaviour of the option “Minimum number of chars for completion”
This was broken by adding the emoji support, after adding the emoji
support it showed only emoji if less characters than “Minimum number
of chars for completion” were typed. But it should show nothing then.

And, if the option “☑ Enable suggestions by Tab key” is used, the
minimum number of characters for completion should be ignored, if the
user types Tab to request a candidate list, this user wish should
always be honored, no matter how many characters were typed.  In this
mode of manually requesting the candidate list by typing Tab, it makes
no sense preventing the candidate list from appearing even if the user
has typed Tab just because the number of characters is smaller then
“Minimum number of chars for completion”.