If hunspell thinks a word is spelled correctly, it must be a match
Valid words might not be found by matching in the word list extracted
from a dictionary.  This is because that word list might not contain
all possible word forms.  When loading the dictionary, the prefix and
suffix information is currently ignored¹. But if a word is spelled
correctly according to hunspell, this must surely be counted as a
match.  Hunspell knows about the prefixes and suffixes.

¹ That may change in future, maybe the “unmunch” utility can be used
to get all possible words. But “unmunch” doesn’t seem to work
correctly for all dictionaries at the moment. And for some
dictionaries, although result produced by “unmunch” seems correct, it
is huge.  /usr/share/myspell/it_IT.dic is only 1.3M, but the unmunched
version is 625M! So I am not sure whether “unmunch” can be used,
maybe only for some dictionaries with some post-processing?