Emphasize words found in the user database correctly
A score of > 1.0 is impossible (except for emoji).  Therefore,
checking user_freq > 10.0 never did anything at all.  Use > 0 instead
of > 10.0 to emphasize words found in the user database correctly.

Do the emphasizing only when DEBUG_LEVEL > 0. Colours for emphasizing
do not work in Gnome3 anyway and they might look very bad in other
desktops depending on the background colour.

Using the textual markers ✓ (for spell checking suggestions) and ⭐
(for suggestions found in the user database) would work on any
desktop.  But for the moment I display these only when DEBUG_LEVEL > 0
as well.  For debugging I find it useful to see suggestions from the
user database marked by ⭐, then I know that they can be deleted with
a key binding or mouse click, those without the ⭐ cannot be deleted.
But this is probably only useful for debugging.