If hunspell suggests a word which differs only by accents from input_phrase, it is a match
input_phrase might not have matched accent insensitively in the list
of words extracted from the dictionary because the list of words might
not contain all possible word forms.

But if hunspell suggests something which differs from input_phrase
only by accents, that means it should have matched if the word list
only had contained all possible word forms.

So this should be counted as a match too.