Make Control+MouseButton1 remove the clicked candidate from the user database
And remove the mouse binding for MouseButton2 to remove the clicked
candidate from the user database. So MouseButton2 does nothing anymore
for the moment.

The key binding for committing a candidate is F1…F9, for removing a
candidate it is Control+F1…Control+F9.

MouseButton1 commits the clicked candidate, so it seems natural to
make Control+MouseButton1 remove the clicked candidate from the user
database. More natural than a different mouse button. Button 2 is now
left for further extensions.