Change key binding for looking up related candidates from Alt+F12 to AltGr+F12
AltGr+<something> is more likely not to conflict with something else.

A keyboard layout supporting AltGr (“ISO_Level3_Shift”) is recommended
anyway, AltGr+Space is already used to input literal spaces into the
preëdit (Alt+Space would conflict with lots of stuff).

Unfortunately that means that using the standard “us” keyboard layout
is not a good choice when using ibus-typing-booster as some key
bindings are not available then. A good alternative to using the “us”
layout is “us(euro)” which adds the € symbol on AltGr+5 but is
otherwise identical the the standard us layout. I.e., on the command line
one would use

    setxkbmap 'us(euro)'

instead of

    setxkbmap us