Make showing of similar emoji work even if emoji preditions are off
Try to find similar emoji even if emoji predictions are turned off.
Even when they are turned off, an emoji might show up in the candidate
list because it was found in the user database. When using the key or
mouse binding to show related candidates on such an emoji, make that
work even if emoji predictions are off. The reason for turning emoji
predictions off is usually higher speed. But if one explicitely
requests similar emoji with a key or mouse binding, this should
always work.

This should *not* turn emoji predictions on though!  The emoji
predictions mode should not change by requesting similar emoji.