If “☑ Enable suggestions by Tab key” option is on, any preëdit change should hide the lookup table
When one requests a lookup table by typing Tab and then deletes
characters from the preëdit (For example by typing Backspace), it is
annoying if the lookup table stays and updates slowly.  Hiding the
lookup table again when anything in the preëdit changes makes editing
the preëdit of course much faster.

The main reasons to use the option “☑ Enable suggestions by Tab key”
are speed and avoiding the flicker by a constantly changing lookup
table.  I.e. it seems to fit better to the spirit of that option to go
for maximum speed and minimum flicker and show an lookup table really
*only* when requested by typing Tab and immediately hide it again
when anything changes in the preëdit.