Load CLDR data for *all* languages in the _expand_languages() list
This is to make Swiss German (de_CH) work. The de_CH.xml file
in CLDR is very small, it contains only those emoji annotations
which contain the letter “ß” in ghe de.xml file. As “ß” is not
used in Switzerland, the de_CH.xml file replaces this with “ss”.
All emoji annotations which do not contain the letter “ß” in
de.xml are missing from de_CH.xml.

Therefore, loading only the de_CH.xml file when instantiating

    EmojiMatcher(languages=['de_CH'])

would be quite bad, most emoji would not be found when typing German
then. It is better to load all .xml files from the _expand_languages()
list, i.e. in case of de_CH load both de_CH.xml and de.xml.

In case of Portuguese, both pt_PT.xml and pt.xml are complete.  So
loading pt_PT.xml only would be enough. But loading pt.xml as well
when pt_PT is requested is probably not bad.

This patch also has the side effect that en.xml is always loaded now,
which is probably a good thing.