If “☑ Enable suggestions by Tab key” is on, don’t autocommit digits
Digits may need to be autocommitted if they are used as selection keys
to select a candidate. Whey they select a candidate, they cannot be
typed into the preëdit anymore. But that problem occurs only when the
candidate list is visible.

But when “☑ Enable suggestions by Tab key” is used, the candidate list
is hidden most of the time anyway. And the it is no problem to type
digits into the preëdit. Having them in the preëdit instead of
autocommitting them is better because then it is possible to complete
something starting with digits.