If “☑ Enable suggestions by Tab key” is on make it possible to close the lookup table with Escape but keep the preëdit
Now, when “☑ Enable suggestions by Tab key” is on, it maybe necessary to
type Escape up to 4 times in a row to reset everything.

For example, when a candidate list showing related emoji or related words
is open and one of the candidates is selected (because the arrow-up/arrow-down
or page-up/page-down keys were used in the candidate list). Then:

1st Escape: Deselects the candidate and shows the first page of the
            candidate list of related candidates
2nd Escape: Goes back to the original candidates list.
3rd Escape: Closes the original candidate list but keeps the preëdit
4th Escape: Clears the preëdit as well.