Down, Up, Page_Down, and Page_Up should trigger a commit and be passed to the application if possible
If a lookup table is shown, Down, Up, Page_Down, and Page_Up have to navigate
the lookup table of course.

And, if more than one input method is used, Control+Down and Control+Up have
to switch the input method used for the preëdit.

But if that is not the case, these keys should be passed to the application.

I.e.:

- If only one input method is used, but a lookup table is shown,
  Control+Down and Control+Up navigate the lookup table, switching
  the input method makes no sense.

- If only one input method is used and a lookup table is *not* shown,
  Control+Down and Control+Up are passed to the application

- If a lookup table is shown, Down, Up, Page_Down, and Page_Up
  navigate the lookup table

- If a lookup table is *not* shown, Down, Up, Page_Down, and Page_Up
  are passed to the application.

This improves the behaviour especially when the option “☑ Enable
suggestions by Tab key” is used. Passing keys to the application when
they are not needed by ibus-typing-booster is always better than
grabbing keys for no reason.