Make triggering a commit with “Left” or “Control+Left” work correctly in “Tab enable mode ” again
This was broken only in “Tab enable mode” (since the improvement to
show the lookup table in “Tab enable mode” *only* on explicit
request). Because when no lookup table was shown and a commit was
triggered by “Left” or “Control+Left”, a candidate was still fetched
from the lookup table, but this candidate was of course empty. Then
that candidate was checked for emptiness and “False” (pass key to the
application) was returned. I.e. the application got only one single
“Left”. But when triggering a commit with “Left” or “Control+Left”,
more than one “Left” has to be passed to the application, one “Left” or
each character in the committed string to get to the left side of the
committed string and then one more “Left” for key which triggered the
commit.