Make the tooltips for the floating panel the same as the labels in the menu
Seems to describe the symbols in the floating panel well enough and
reduces the number of messages to translate by two. I also found
Switch the “Unicode symbols and emoji prediction” mode awkward to
translate, just “Unicode symbols and emoji predictions” seems better,
the button in the panel pops up a On/Off menu, so it is obvious anyway
that this is a switch.

And change “Off the record” -> “Off the record mode” in the setup tool

Reduces the number of strings to translate by one.