Make the list of characters to auto commit configurable
Empty list by default.

The characters in this list automatically commit the preëdit and add a
space.  For example, if “.” is an autocommitting character, this saves
you typing a space manually after the end of a sentence. You can
freely edit this list, a reasonable value might be “.,;:?!)”. You
should not add characters to that list which are needed by your input
method, for example if you use Latin-Pre (t-latn-pre) it would be a
bad idea to add “.” to that list because it would prevent you from
typing “.s” to get “ṡ”. You can also disable this feature completely
by making the list empty.

Even if a character is in that list, typing that character might still
not autocommit if it is sucked up by an input method. For example, if
“,” is on that list and one is using Latin-Post and types “a,”, the
input method converts this to “ą” and it would not be nice to
autocommit then. Therefore, an auto commit only happens if the last
character of the preëdit is still the auto commit character after
transliteration has been done.

Also, an auto commit is prevented if the preëdit does not contain at
least one letter type character in addition to the auto commit
character. For example, if one types “:-)” and “)” is an auto commit
character, auto commit does not happen because in that case one
probably wanted to match the emoji “:-)” -> “🙂”.