Make it possible to use a database in different locations than the default
Also make the user of in a memory database possible.

This is mainly for the unit tests. When running the unit tests, it is
not good to use the database in the default location because a
database might already exist there. And then the results of the unit
tests may depend on what is already in that database. And the unit
tests should not change a database which is used for real user input.

For the unit tests, a new, empty database should be used, either in memory
or in a temporary directory.