Trigger emoji lookup when the input starts or ends with '_' or ' '
The emoji search is a bit slow. Therefore, one might want to set the
emoji mode to off. But nevertheless one might want to input an emoji
from time to time. With this change, one does not need to switch emoji
mode on and off again to quickly input one emoji, it is enough to type
a search key starting or ending with '_' or ' '. If one only rarely
inputs emoji, this is probably the fastest way to do it.