Return immediately if _update_candidates() is called with empty input
Resolves: rhbz#1413082 - When emoji mode is off *and* the option “Enable suggestions by Tab key” is off *and* one types very fast, ibus-typing-booster sometimes crashes
See: https://bugzilla.redhat.com/show_bug.cgi?id=1413082

Since I triggered emoji when the input starts or ends with '_' or ' ',
I sometimes experienced ibus-typing-booster failing with this backtrace:

  File "/local/mfabian/src/ibus-typing-booster/engine/hunspell_table.py", line 685, in _update_candidates
    or self._typed_string[0] in (' ', '_')
IndexError: list index out of range

So sometimes _update_candidates() was apparently called with empty
input.  This happened only when emoji mode was off *and* the option
“Enable suggestions by Tab key” was off *and* I typed very fast. The
reason is probably that _update_candidates() is usually called using
GLib.idle_add(). When one types a word very fast followed by a space to
commit, and then hesitates a bit, the first idle time occurs when
the input is empty again because a commit just happened.  Of course
_update_candidates() should not fail in this case, when the input is
empty, the best _update_candidates() can do is clear the lookup table
and return immediately.