Calculate the maximum word length for each dictionary individually
It was harcoded to 40 which was about correct for the German hunspell
dictionary.  But some dictionaries have longer words (it_IT has a
maximum word length of 48) and some have shorter words (en_GB has a
maximum word length of 28). So using the hardcoded value of 40 might
miss some matches in the Italien dictionary and waste some time using
the British English dictionary trying to complete words which are
already longer then the maximum word length in the en_GB dictionary.