Handle Return and Enter correctly when the cursor is not at the end of the preëdit
Resolves: rhbz#1418313 - Committing a preëdit with Enter or Return behaves badly when the cursor is not at the end of the preëdit
See: https://bugzilla.redhat.com/show_bug.cgi?id=1418313

Until now, committing a preëdit with Return or Enter did commit the
complete preëdit and forward the Return or Enter to the application
afterwards. For example, if the preëdit looks like “wo|rd” (where “|”
marks the positon of the cursor), committing with Return did send
“word” + Return the application. But sending “wo” + Return + “rd”
to the application is much more natural behaviour.

This commit tries to achieve this.