Don’t sort the labels when listing similar emoji
This is mostly to get the kanji and their hiragana for Japanese and
the Chinese characters and their pinyin for Chinese listed in the
right order. I.e. Chinese characters first immediately followed by
their pronunciation in hiragana or pinyin.

I.e. when using Chinese, the listing of similar emoji should look like this:

    >>> matcher = EmojiMatcher(languages = ['zh_CN'])
    >>> matcher.similar('🏇', match_limit=1)
    [('🏇', "赛马 ['赛马', 'sàimǎ', '马', 'mǎ']", 4)]

and not like this:

    >>> matcher = EmojiMatcher(languages = ['zh_CN'])
    >>> matcher.similar('🏇', match_limit=1)
    [('🏇', "赛马 ['mǎ', 'sàimǎ', '赛马', '马']", 4)]