Fix duplicate listing of labels when looking up similar emoji
When using EmojiMatcher was used which had a language list which
contained the same language twice, the labels (keywords, categories)
of similar emoji were counted twice. Having the same language twice in
the list can happen by accident because of fallbacks. Here we have 'es'
twice in the expanded list of languages:

>>> itb_emoji._expand_languages(['es_ES',  'it_IT', 'es_MX', 'de_DE', 'en_US', 'ja_JP'])
['es_ES', 'es', 'it_IT', 'it', 'es_MX', 'es_419', 'es', 'de_DE', 'de', 'en_US', 'en_001', 'en', 'ja_JP', 'ja']

In that case, the shared labels of similar emoji were counted twice for 'es'.

This commit fixes that and counts them only once even in that case.