Nicer display of the matching labels when looking up similar emoji
Old:

    >>> matcher.similar('🏄‍♂', match_limit = 2)
    [('🏄\u200d♂', "hombre haciendo surf ['🏄\\u200d♂', 'hombre', 'surf', 'surfista']", 4), ('🏄\u200d♀', "mujer haciendo surf ['surf', 'surfista']", 2)]

Here, the first label of the most similar emoji displayed as “🏄\u200d♂”
in the lookup table, which looks very ugly.

New:

    >>> matcher.similar('🏄‍♂', match_limit = 2)
    [('🏄‍\u200d♂', 'hombre haciendo surf [🏄\u200d‍♂, hombre, surf, surfista]', 4), ('🏄\u200d‍♀', 'mujer haciendo surf [surf, surfista]', 2)]

Now the first label of the most similar emoji displays as “🏄‍♂”, which
looks much better. On top of that, I removed the single quotes around
the matching labels. This makes the list of matching labels display a
bit shorter in the lookup table. It is nicer if the lookup table is
not wider than necessary.