Set WM_CLASS of emoji-picker and ibus-setup-typing-booster correctly
This is needed to get the application name and icon shown correctly by
gnome-shell in the top bar.

gnome-shell seems to use the first argument of set_wmclass() to find
the .desktop file.  If the .desktop file can be found, the name shown
by gnome-shell in the top bar comes from that .desktop file and the
icon to show is also read from that .desktop file. If the .desktop
file cannot be found, the second argument of set_wmclass() is shown by
gnome-shell in the top bar.

Therefore, for emoji-picker we need:

    set_wmclass('emoji-picker', 'Emoji Picker')

and for ibus-setup-typing-booster we need:

    set_wmclass('ibus-setup-typing-booster', 'Typing Booster Preferences')