Set the emoji font only for the emoji, not for its name in the search results
When typing “cat” into the search entry, one of the results is:

🐱 cat face

🐱 is the emoji and “cat face” its name. Better not change the font for the
name to the emoji font, it is better to leave it alone. Usually
emoji fonts don’t have glyphs for letters, but one never knows.

There is at least one case where setting the emoji font for the name
as well looks ugly:

When typing something like “2019” into the search entry, one gets the result:

’ U+2019 RIGHT SINGLE QUOTATION MARK

The ’ is the “emoji” and “U+2019 RIGHT SINGLE QUOTATION MARK” is its name
or description. Setting the emoji font for “U+2019 RIGHT SINGLE QUOTATION MARK”
may look very ugly because some emoji fonts have funny glyphs for digits,
so the digits in “U+2019” may look very strange. For example the
“Noto color emoji” has weird looking glyphs for digits.