Make skin tone popover scrollable and limit its maximum size
In most cases, this does not change anything if there are only 6 emoji
in the popover because a popover for only 6 emoji usually fits into
the main window without scrolling. But it helps in cornercases when
the main window has a very small size. And it might be useful if one
wants to display more than 6 emoji in such a popover.