Skip the emoji which already have skin tone modifiers in itb_emoji.emoji_by_label()
This is to make the real number of emoji listed shown when browsing
the categories (In the treeview at the left and in the header bar).

The skin tone variants are generated as needed when long pressing on a
base emoji which supports skin tones. They are not listed directly
when browsing the categories. Therefore, it is better to filter these
out already in itb_emoji.emoji_by_label() instead of doing it when
filling the flow box. This way the counts displayed are correct.