Make it possible to limit to a certain Unicode version range when browsing emoji categories
The search feature will still  find emoji of all Unicode versions,
but when browsing the emoji by category, the results shown can be limited
to a certain range of Unicode versions.

For example:

    emoji-picker --emoji_unicode_min 8.0 --emoji_unicode_max 9.0

will show only emoji which were added between Unicode 8.0 and 9.0 (inclusive)
when browsing the emoji by category.