Add an option whether to use pango font fallback
- Extend itb_pango.get_fonts_used_for_text() to support an option
  whether to use pango font fallback.
- Add a command line option to choose whether to use pango font fallback.
- Add a check button to choose whether to use pango font fallback.
- Save the current value of the pango font fallback option in the options file.

Disabling font fallback can be quite useful to see easily which glyphs
are missing in a certain font.

On top of that, with the latest fontconfig and pango it is impossible
to choose any other emoji font than the default emoji font according
to the fontconfig setup as long as the pango font fallback is enabled.
I.e. the font selection menu in emoji-picker does not work at all
as long as the pango font fallback is enabled.

Disabling the pango font fallback makes it at least possible to select
other color emoji fonts if the fontconfig default is to use color for
emoji or other black and white emoji fonts if the fontconfig default
is to use black and white for emoji.

- Improve font selection by listing the best emoji fonts on top
- Add generic “emoji” font family to font selection