Move the buttons to learn from a file and to delete learned data to the options tab
Having an extra tab for only these two buttons is a bit too much, it makes
the window unnecessarily wide.

Three lines were removed from the options tab because of merging
the engines (The “Add direct input” check box, the input method
combo box and the input method help button are not in the options
tab anmore). So adding two more buttons to the options tab should
not make that too tall.