Use the rocket icon emoji_u1f680.svg from the “Noto Color Emoji” font
When using the Gnome desktop, the input method switcher shows text and
for ibus-typing booster “🚀 U+1F680 ROCKET” is used. By default this
usually shows the rocket emoji from the “Noto Color Emoji” font because
that font is typically used by default to show emoji.

But on non-Gnome desktops, an icon is shown. The old
ibus-typing-booster.svg icon shows a globe in front of a keyboard. So
it looks very different from what is shown in Gnome. That is not nice,
it is better if it looks similar on all desktops, therefore it is
nicer to show an icon which looks like the emoji if a emoji cannot be
shown as text.

Therefore, this commit changes the icon to the rocket icon
emoji_u1f680.svg from the “Noto Color Emoji” font.