Better Tab handling, use Tab to switch to the next candidate, not to commit
Suggested by Fujiwara Takao. Switching to the next candidate when
typing Tab is more useful than committing and sending a Tab. The Tab
key is in a very ergonomic position and when using the option “Enable
suggestions by Tab key”, the finger is on the Tab key already anyway
when the candidate list pops up. So it is easier to just continue
pressing Tab until the right candidate is reached than moving the
finger to a number key to select a candidate. The advantage of using
Tab to select a candidate is even greater if the option “Use digits as
select keys” is disabled, the F1-F9 keys are quite far a way from the
fingers. If one really wants to commit and send a Tab to the application,
one can still commit by typing the right arrow key and then Tab.