Do not try to commit if index is >= the current number of candidates
Resolves: rhbz#1630349 - On selecting candidate option number, greater than suggested number of candidates (keyboard shortcuts), typing-booster stops working.
See: https://bugzilla.redhat.com/show_bug.cgi?id=1630349

The page size returned by self._lookup_table.get_page_size() is
constant, it does not change with the number of candidates displayed
in the current page.

Therefore, numer of candidates displayed in the current page
may be smaller than self._lookup_table.get_page_size().

If a selection key 1-9 or F1-F9 is pressed to commit a candidat with
that label, it is not enough to check whether the selected index is
smaller than the page size, one must also check whether the index
is beyond the last candidate currently displayed.