Remove code to check whether another instance of the setup tool is running.
I think there is no need to check this, it works fine to run multiple
setup tools. If one setup tool window is already there but invisible
because it is buried under other windows, why is it useful to get a
dialog box saying that another tool is running already and then exit?
Why not just open another setup tool, instead of searching where the
other one is?