Commit the current preëdit when the focus changes
Commit the current preëdit when the focus is moved to another window
or when the focus is moved using the mouse in the current window.

The commit is achieved by using

    update_preedit_text_with_mode(,,,IBus.PreeditFocusMode.COMMIT)

instead of

    update_preedit_text(,,,)

If the cursor has been moved with the mouse in the current window,
IBus emits 'reset'. Therefore, a do_reset() method is added to handle
that signal.

Both in do_reset() and in do_focus_out() code is added to record
what has been in the already committed preëdit to the user database.