When preëdit empty or lookup table not enabled by key: clear the lookup table
When making the preëdit empty by typing backspace and the lookup table
is only hidden but not cleared, the lookup table may still contain
candidates although the preëdit is empty. Then, if a key to move
around in the lookup table is typed (like Up for
'select_next_candidate'), this key causes moving around in the
invisible lookup table instead of being passed to the application as
desired.

Also, when the preëdit is not empty but the lookup table is hidden
because it is not enabled by the key to enable the lookup, hotkeys
which would move around in the lookup table if it were visible should
trigger a commit if the lookup table is invisible. The easiest way to
achieve this is to not only hide the lookup table but also clear it.
Then, when a hotkey for moving around in the lookup table matches, one
can skip that hotkey action if the lookup table is empty.