Guess current Unicode version to set --emoji_unicode_max in emoji_picker
When the Unicode data for emoji is updated to a draft version newer
than the current released version of Unicode and the
--emoji_unicode_max option of emoji-picker does not limit the display
to the current released version of Unicode, it is almost impossible
that a reasonable font for the new emoji is already available. Then it
is practically guaranteed that one will see a lot of replacement
characters. Updating the emoji data to a draft version can be useful
for testing emoji-picker, but for normal users it probably makes no
sense to display the draft emoji before the respective Unicode version
is released. It might even take quite a while after a Unicode release
until suitable fonts become available. So for normal users it is better
to limit the display to the currently released versions of Unicode.
For testing one can use the --emoji_unicode_max option to set a higher
Unicode version.