Key bindings should match independent of whether Num Lock or Caps Lock are on or off.
For example, the default key bindings to select the next candidate
are

    Tab, ISO_Left_Tab, Down, KP_Down

These should also match when Num Lock or Caps Lock (or both) are on,
otherwise ibus-typing-booster becomes unusable when Num Lock or Caps
Lock are on because Tab would then commit instead of selecting the
next candidate.

Theoretically, one could also add

    Lock+Tab, Mod2+Tab, Lock+Mod2+Tab

to the default list of keys to select the next candidate.

But then one would always need to add all these Lock and Mod2 variations
when editing key bindings. This would be very confusing and I don’t think
this flexibility would add any real value.

It is better to just not use Lock (Caps Lock) and Mod2 (Num Lock)
in key bindings.