Only show the label in the panel menus for emoji-mode and off-the-record-mode
I.e. only show

    “Unicode symbols and emoji predictions”
    “Off the record mode”

instead of

    “☑🙂 Unicode symbols and emoji predictions”
    “☑🕵 Off the record mode”

“☑🙂” and “☑🕵” are still there in the small floating panel on
non-Gnome desktops.  There it makes sense to display it with emoji to
show the information with minmal use of space. In the Gnome-panel and
other panel menus, it is a bit wasteful and maybe even ugly. Seeing
the on/off status there directly seems not necessary.