Move “Dictionaries and input methods” Tab in the setup tool to the first position
... and select it by default when the setup tool starts.

First time users should see this first. ibus-typing-booster guesses
the dictionaries and input methods for first time users from the locale.
But that can easily be wrong. For example if the user runs in an
English locale but his native language is not English. In that case
he surely wants to add a dictionary for his native language here and
a suitable input method.

Even if one already uses ibus-typing-booster for some time, this tab
still seems to be the most often used one, at least I use it most
often to add or remove languages. I change the other options of
ibus-typing-booster much less frequently, after using it for a while I
know which options I like and which I don’t like.

But adding or removing languages continues to be useful for me even
after long time use. I sometimes type stuff in English, German,
French, Spanisch, and Italian. ibus-typing-booster supports using
dictionaries for many languages at the same time, but using more
dictionaries at the same time makes it slower and lowers the
prediction quality at least somewhat. Therefore, if I type only only
English and German for a while, it may make sense that I remove the
French, Spanish, and Italian dictionaries and add them back later when
I need them.

So I think this tab is the most often used one even for longtime users
and thus should go first and be selected at startup of the setup tool.