Add new command 'toggle_hide_input' to hide the preëdit, lookup table, and auxiliary text
Suggested by Takao Fujiwara. Might be useful to hide quickly what one
is typing. For example when typing at a password prompt in a terminal,
this might be useful.

While in “hidden mode”, the input is also not stored into the user database,
just like with “off the record mode”.