Add direct input mode, i.e. a key binding to toggle ibus-typing-booster On/Off
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/25

This key binding is empty by default.

If this key binding is set to something non-empty, an additional menu
appears in the desktop panel and in the floating panel to switch the
input mode beween “On” (ibus-typing-booster) and “Off” (direct input).
I.e. with an non-empty key binding one can switch between the input modes
not only with the key binding but also via the panels and the panels
show the current state of the input mode.

If the key binding is made empty, the extra menu entry disappears.  So
this extra key binding option should have no disadvantages for users
who do not need a direct input mode, they can just leave the key
binding empty.