Don’t reinitialize the dictionaries if only the order of the dictionaries has changed
If changing the order of the dictionaries triggers a reinitialization
of the dictionaries, it causes a delay until ibus-typing-booster is
ready again to accept input.

When one uses typing to input, one does not really need to change the
order of dictionaries often. The priority of the dictionaries has
only a small influence on the prediction quality.

But when using speech recognition, the order of the dictionaries is
very important. Speech recognition is done for the language of the
dictionary with the highest priority. So every time one wants to
switch the language for speech recognition, one needs to change the
priority order of the dictionaries.

If only the order of the dictionaries is changed, there is no need for
the slow reinitialization. Just change the order of the dictionary
objects, this is very fast and ibus-typing-booster is ready for input
again almost immediately.

I.e. switching the language for speech recognition is very fast now.